requirements.txt
anyio==3.6.2
APScheduler==3.6.3
async-generator==1.10
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-telegram-bot==13.15
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tornado==6.1
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
wsproto==1.2.0
